# server/api-spec/openapi.yaml
openapi: 3.0.0
info:
  title: Stories API
  description: API for managing stories in the application.
  version: "1.0.0"
servers:
- url: http://ec2-54-196-0-54.compute-1.amazonaws.com:3001/v1
  description: Development server
- url: http://localhost:3001/v1
  description: Local server
paths:
...
  /twyne/story/{storyId}:
    get:
      tags:
      - Twyne
      summary: List Twynes by Story ID
      description: This endpoint returns all Twynes associated with a given Story ID
      operationId: listByStoryId
      parameters:
      - name: storyId
        in: path
        description: ID of the Story to find Twynes for
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Twynes found successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Twyne'
        '400':
          description: Bad request
        '404':
          description: Twynes not found
        '500':
          description: Server error
  /twyne:
    post:
      tags:
      - Twynes
      summary: Create a new Twyne
      description: This endpoint allows for the creation of a new Twyne
      operationId: createTwyne
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Twyne'
      responses:
        '200':
          description: Twyne created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Twyne'
        '400':
          description: Bad request
        '500':
          description: Server error
  /twyne/{id}:
    delete:
      tags:
      - Twynes
      summary: Delete a Twyne
      description: This endpoint allows for the deletion of a Twyne
      operationId: deleteTwyne
      parameters:
      - name: id
        in: path
        description: ID of the Twyne to delete
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Twyne deleted successfully
        '400':
          description: Bad request
        '404':
          description: Twyne not found
        '500':
          description: Server error
    get:
      tags:
      - Twynes
      summary: Find a Twyne by ID
      description: This endpoint returns a Twyne based on its ID
      operationId: findById
      parameters:
      - name: id
        in: path
        description: ID of the Twyne to find
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Twyne found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Twyne'
        '400':
          description: Bad request
        '404':
          description: Twyne not found
        '500':
          description: Server error
    patch:
      tags:
      - Twynes
      summary: Update a Twyne
      description: This endpoint allows for the updating of a Twyne
      operationId: updateTwyne
      parameters:
      - name: id
        in: path
        description: ID of the Twyne to update
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Twyne'
      responses:
        '200':
          description: Twyne updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Twyne'
        '400':
          description: Bad request
        '404':
          description: Twyne not found
        '500':
          description: Server error
...
  /users:
    post:
      tags:
      - Users
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags:
      - Users
      summary: Update a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:

        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{userId}:
    parameters:
    - in: path
      name: userId
      required: true
      schema:
        type: string
    get:
      tags:
      - Users
      summary: Find a single user
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    delete:
      tags:
      - Users
      summary: Delete a user
      responses:
        '204':
          description: User deleted successfully
components:
  securitySchemes:
    BearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, arbitrary value for
  schemas:
...
    Twyne:
      type: object
      properties:
        _id:
          type: string
          format: ObjectId
          description: The unique identifier for the Twyne
        storyId:
          type: string
          format: ObjectId
          description: The identifier for the story
        prompts:
          type: array
          items:
            type: string
            format: ObjectId
          description: The identifiers for the prompts
        owner:
          type: string
          format: ObjectId
          description: The identifier for the owner
        coCreators:
          type: array
          items:
            type: string
            format: ObjectId
          description: The identifiers for the co-creators
        contributors:
          type: array
          items:
            type: string
            format: ObjectId
          description: The identifiers for the contributors
        storyline:
          type: string
          format: ObjectId
          description: The identifier for the storyline
        edit:
          type: string
          format: ObjectId
          description: The identifier for the edit
        storylineTemplate:
          type: string
          format: ObjectId
          description: The identifier for the storyline template
...
    UserRegistration:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required:
      - username
      - email
      - password
    User:
      type: object
      properties:
        _id:
          type: string
        username:
          type: string
          description: The user's username
          example: "exampleUser"
        email:
          type: string
          description: The user's email
          example: "exampleUser@example.com"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        profile:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            avatarUrl:
              type: string
            bio:
              type: string
        settings:
          type: object
          properties:
            notifications:
              type: object
              properties:
                email:
                  type: boolean
                push:
                  type: boolean
        status:
          type: string
        roles:
          type: array
          items:
            type: string
      required:
      - username
      - email
      - salt
...
security:
- BearerAuth: []